mcastcapture: fix more problems with missing libcurl.
On machines without libcurl-dev installed (such as the autobuilder machine),
host-google_mcastcapture couldn't build.  Previous fix didn't work,
unfortunately.  Tested this one by uninstalling libcurl-dev on my machine.

Also added some LD_LIBRARY_PATH at the toplevel so it will use the libcurl
from $(HOSTDIR) instead of the host's built-in copy.  This is harder to
test, but I confirmed it with ldd and it seems to work.

Change-Id: I41800515e7681030b97fd84557d94f5102bdfb55