Hackarama to make google_platform-test work
The test build is sort of weird, it builds the code using the host
compiler (as far as I can tell) instead of the target.  The target
version of the proto build needs headers from the protobuf build output
so this just passes in TARGETDIR set to the host path instead.

Change-Id: I2c68409356e217611f63ed1f6042193a9674a684