Increase maximum allowed value for UDP socket buffer size.
Currently rmem_max is set to 512KB, which is actually half of what we
try to set in sagesrv (the value is doubled internally so it is still
close to what we want). Increase the limit here to something more
appropriate that gives us more headroom to work with. Note that this
is just an upper boundary, not the actualy buffer size used.

Change-Id: I0bbabbcee87884353c9923c06cba52754f93122f