tlsdate: fix bad behaviour if some IP addresses time out.
When connecting to multiple IP addresses in series, connect() was left to
wait as long as it wanted, which on Linux can be more than a minute.  On
networks that advertise working IPv6 but don't actually have working IPv6,
this could cause initial boot to take more than a minute.

To compound the problem, the RTT calculation uses the start and end times
for the entire process, including all the IP addresses tlsdate-helper tries
to connect to.  So if one address timed out and the second one worked, the
total RTT would be calculated as (timeout + real rtt), which is way too
large, and the clock compensation would make things much worse.

To fix this, connect to all IP addresses for the given hostname in parallel
and use the first one that works.

While we're here, enable tlsdate verbose mode in run-tlsdate.  Log messages
are cheap, but debugging is hard.

Change-Id: I968eced141f2af3708cbc938f2cda5d61926fd73