run-tlsdate: fix clockspeed startup race condition.
The clockspeed daemon would die repeatedly if we managed to try to adjust
the clock before clockspeed had finished starting (and creating its fifo).

This probably started happening only after the recent optimization to pin
the NTP server as long as it is responding.  That avoided the need for DNS
lookups before doing the first clock adjustments, so it suddenly was able to
happen very fast.

Change-Id: Iba2f51ddca185d03e4e5a9b18df5ac2b4f53156e