Change how GFLT builds the final uImage
Instead of using the kernel to build a uImage, build a zImage and then
have the ginstall.mk  create the uImage.  This allows changing the
version string to something custom, and is also much closer to the gftv
build system, so merging the two ginstall.mk files should be a lot
easier now.

Also:
Regenerate gflt110_defconfig with 'make oldconfig'
Regenerate gflt200_defconfig with 'make oldconfig'

Change-Id: I6ca5efeb469c6407f1f06153da894c47a1f92d16