mac80211: disable setting QoS flags from DSCP values.
This causes some problems with the current ath10k driver, which seems to do
wrong things with ADDBA if multiple QoS TIDs are in use at once.  In any
case, it's not clear that the default settings were what we wanted, so this
shouldn't hurt performance for now.  Eventually we can remove this when the
ath10k driver has been fixed up.

Fixes b/13332693.

Change-Id: Ie979907fc8faf9f2070989a1b7adaeb861d12507