Only test hnvram as part of host-google_platform-test
All other tests are already run as part of google_platform-test
(GOOGLE_PLATFORM_TEST_CMDS). The google_platform package does not follow
buildroot conventions in the sense that it builds binaries for the
target *and* the host in build/google_platform-HEAD/. For example, you
find a usleep for the target architecture and a host-usleep for the host
in the cmd subdirectory under build/google_platform-HEAD/. This is true
for most subdirectories with the notable exception of hnvram which only
builds a binary for the target and not the host.

Change-Id: I8459cd88cbc0feeef80d12bd2dd0066ff81fe10e