openssl: remove leading / from relative libdir path
libdir is a relative path which is appended to a base path in the Makefile.
The Makefile adds a separating / when composing the absolute path, therefore
the leading / is not needed

While the additional slash isn't an issue for openssl itself it can cause
problems when cross-compiling other packages which (directly or indirectly)
depend on shared libraries of this package. The path with the double slash
is put into <library>.pc. This file is read by autotools when building dependant
packages to obtain information about the library, like the flags to be passed to
the linker. When cross compiling libtool normally transforms an expression like
"-L/usr/lib" to "-L<sysroot>/usr/lib" .This is not the case for "-L/usr//lib".
In consequence the linker tries to link a target binary against libraries of
the host system, if they are present in the directory passed to the linker

Change-Id: I216badcfdc9e006a96a9e0cb90b632ba9eb4b252