Work around multicast forwarding misbehavior.
Hosts are not supposed to have to subscribe to 224.0.0.1,
the all-hosts address. Multicast switches are supposed to
flood it to all ports.

However there are IGMP snooping switches out there which do
not reliably do so. If any switch port subscribes to a multicast
address which happens to map to the same Ethernet MAC address
as 224.0.0.1, they stop flooding it to all ports and only
forward that MAC to the ports where they think there is a group
member.

To work around it, deliberately subscribe to a multicast group
which maps to 01:00:5E:00:00:01. That way, 224.0.0.1 will
always be forwarded to Storage Box.

Change-Id: I317ff5502beeab60e50e3c3f8cad753de793920b