Rearrange CPU affinity settings on all CPE 1.0 and 2.0 devices.
Main changes:
- CPU2 instead of CPU1 is now the realtime CPU (so that unspecified IRQs
  default to CPU1, the non-realtime CPU)
- Make sure SATA is not on the realtime CPU

Not changed:
- sagesrv should maybe be locked to the realtime CPU instead of
  free-floating.  (If a SATA IRQ is slow, it could interrupt a sagesrv thread
  on the non-realtime CPU for an undefined amount of time, and cause data
  loss on that stream.  Other sagesrv threads could run on the other CPU,
  but I believe this one would be stuck until the IRQ handler returns.)

However, sagesrv is pretty cpu-intensive and may not be able to do all its
work on a single CPU core right now.  We can change that later.  Another
option would be to move only its recording threads (which are sensitive to
multicast loss) onto the realtime CPU, and let the others float freely.

Change-Id: I23d382c258f7089706d8e17cb2c596f111e57fe7