Enabled progressive deletion of files in 128MB increments
Testing shows this alleviates disk slowdown when file deletions occur
and the 128MB number was chosen because that matches the block group
size used in ext4. This may end up deleting more than one extent at a
time; but it also ensures we always delete AT LEAST one extent at a time
so we are always making progress on deletions that don't involve file
size reduction along with writes to extents we will have to redo again.

Change-Id: Ib711358c73c29c1e580d793b11f114959206821e