hostapd: remove -d option, change some logs from debug to info.
"hostapd -d" is just too verbose, but by default, without it, hostapd is not
verbose enough.  Change a bunch of log levels to MSG_INFO instead of
MSG_DEBUG, and drop the debug flag.  Hopefully this will capture the
important parts without spamming the logs with tons of irrelevant bits.

Change-Id: Iaa667864c4c0e6f69ebaf952687876df7ae35faf