Copy verity data to tmpfs and create loopback.
If the verity data is on an mtdblock device, which for us means the
verity data is in nand, copy the data out with nanddump, which is bad
block aware, and create a loopback device.  verity isn't bad block aware
and this cause devices to fail if there are bad blocks in flash.

Change-Id: Ie67b3083a7e1980415bdfd4b16fec5c43fad9eb8