backports: mwifiex: eliminate a crash on hostapd shutdown.
The way sta_list_spinlock is used is highly suspect and there are definitely
race conditions.  mwifiex_cfg80211_del_station is called from a netlink
callback at the same time as mwifiex_del_sta_entry is called from the SDIO
thread, and they both try to delete all stations, but in different ways.

The net result is they could end up both modifying sta_list at the same time
and causing a crash.  This patch fixes the one particular crash case we see,
by referring to a station by its MAC address instead of a pointer to its
station structure.  That way, when a station is deleted by surprise in
another thread, we just have a use-after-free bug on the MAC address array,
which quietly fails a subsequent lookup of the node structure but at least
doesn't crash.

The remaining symptom of the race condition is that stations are not
actually deauthorized at hostapd shutdown (because one of the code paths
does not send the firmware a deauth request), which is surprising.  They
are, however, deauthorized when hostapd restarts, so the problem isn't too
serious.

A better fix would probably involve totally rethinking the locking in
mwifiex.  There is little need for it to use such fine-grained locking,
especially around administrative structures like the station list.

b/16846245

Change-Id: Ie4a7921fa1b666803d59585d457d0975556bf7de