ath10k: call userspace ath10k-reset program on firmware crash.
This is the nuclear option: if the firmware crashes, we want to do a cold
reset of the chip.  But there are bugs in the cold reset of the chip, and
warm reset doesn't always work, and the driver also gets into a confused state
(disagreement with the firmware about which things are where) that just
makes things worse sometime later.

Instead, call a script that will unload the driver, reset the PCIe bus
entirely, and reload it again.

You can test it with this command:
  echo hard >/sys/kernel/debug/ieee80211/*/ath10k/simulate_fw_crash

Change-Id: Iad7c61698a5aee2bffd7b7a0d1c97369e7937734