ath10k-reset: more robustness.
UMH_NO_WAIT seems to not actually run the requested program sometimes; based
on a separate test I wrote, this is not actually related to running it in
interrupt/atomic context.  It just sometimes doesn't work and doesn't report
any reason why not.  Rather than trying to debug that right now, switch to
using UMH_WAIT_EXEC instead, which seems to work reliably, but doesn't work
in atomic contexts.  So when the firmware crashes, arrange for the script to
get launched from the ath10k tasklet instead, which can safely use
UMH_WAIT_EXEC.

While we're here, improve ath10k-reset to work even if some of the modules
have previously been removed.

Also, there was a bug in the debugfs handling of simulate_fw_crash, where it
would report the wrong number of bytes being written.  This would sometimes
cause the driver to try to double-reset when simulating a crash, and
hilarity would ensue.

Change-Id: I51de304fb060a8ccd234d5295aa4e9f1f32c39f8