Updated logs upload script to call new variant
It only does this when the kernel is 3.16 or newer on the box
since reading from /dev/kmsg isn't supported on older kernels.

I also changed S01logging to only call logmark-once instead of
calling logmark repeatedly in the loop. The log uploader (old
and new) both do logmark-once themselves, and the new one has
better time accuracy. So its redundant to do it continually
in a loop since that information is only used after the logs are
uploaded.

Change-Id: I2b5aaa7ddc0568c15abadd68ef01d6568e84e4da