Don't try to download packages that can't be downloaded.
Remove sources.buildroot.net as a fallback location for downloading
packages.  This is never the right thing to do.

At some point there was a false interpretation that setting
${PACKAGE}_SOURCE="null" would prevent the download.  If
${PACKAGE}_SOURCE is not NULL then it checks the method to use for
download.  If ${PACKAGE}_SITE_METHOD is equal to "null" then it
exits without attempting a download.

Unfortunately just unsetting ${PACKAGE}_SITE isn't enough to fix
this.  If $BR2_PRIMARY_SITE and $BR2_BACKUP_SITE and ${PACKAGE}_SITE
are all undefined it defaults to dl.sourceforge.net.  I wasn't able to
find how this actually happens but it's documented in
buildroot/docs/manual/adding-packages-gentargets.txt.

Sets ${PACKAGE}_SITE_METHOD=null so when it tries to download from
sourceforge it just exits without an error.

Change-Id: I037c5a09a0fc34474954b74c6d199858f8ed9702