Actually kill the './start' script in S99oregano (b/19363748)
See the bug for details, but the gist is that S99oregano was not actually
killing './start', and instead relying on its child browser_shell to
terminate quickly and with a return code of 0. When this did not happen,
'./start' would return 137 and its (still alive) babysitter would revive
it, to ugly consequences (Oregano would constantly come up and then fail to
get Nexus, crash, and keep trying. Forever.)

Change-Id: If39e1ef3ddbd7401dd7e85266fd0b1ac70dabbc2