buffet: only mount chroot folders once.
I added a mount_once function that checks that if the mountpoint has
already been mounted. Before this cl, every time the script was
called the folders were mounted. This didn't introduce any noticeable
bug, but those folder appeared duplicated in the mount command (and we
had to unmount them multiple times to get rid of the mount points).

Change-Id: Ibb5b51a657dbbde7bc5a89c704a02a509285098e