SpaceCast: fix for lan0 not up after ginstall
It was seen lan0 not up about 1 in 300 upgrades. There is
no error shown in the logs. After reading codes, I found
there is a similar problem in Optimus with br0 in this
version of linux. The fix for br0 is to add one extra call
to set the link up. I added an extra call here. The problem
has not been seen since. So I will submit this change as fix
for this problem.

Google-Bug-Id: 20834837
Change-Id: I1116288d68e57868aa56f18b281bca808b791abb