spacecast: Change CHROMEOS_VERSION variable name
variables defined in the .mk files are global variables, and there was a
collision between the cryptohome and buffet CHROMEOS_VERSION
variable.

This CL Changes both variable names, prepending the respective module
name (CRYPTOHOME and BUFFET) to the variable name. This avoids the
naming collision.

This wouldn't be a problem in practice, because we should use the same
version of both modules, but it is a problem for developing and testing
new versions of a module without changing the other.

Change-Id: Iaf6dc88b50cbe92b0f6b43a1786778fed7921b79