Cast-on-CPE: Add temporary dummy 'logwrapper' script to /bin/
Currently, v2mirroring requires a '/bin/logwrapper' binary to exist for
it to start. During Cast-on-CPE development we mount the Cast binary
files, however this requires a target file to exist. Hence, this dummy
script.

This placeholder script will be either A: replaced when we start putting
cast_shell into images, or B: removed when Cast remove the requirement
to have a /bin/logwrapper.

Change-Id: I38a966935921a8985844655b310b6c0d2c651e45