Split gfsc100 into a dev and prod config
CL ce92c82e introduces a problem with savedefconfig on gfsc100 prod
builds. It disables BR2_PACKAGE_GOOGLE_TEST which was set when the
config file was created. In consequence dependencies which are
implicitly enabled by BR2_PACKAGE_GOOGLE_TEST and needed by non-test
packages are not enabled on prod builds.

Having different configs for prod and dev adds a maintanence overhead
of keeping them in sync, on the plus side no hacks (with potentially
subtle errors) are needed to make a single config file work for both
targets.

Change-Id: Ifc2e421b38ca440a00ff7487a7a2656cb8aa7657