dash: fix signal names on mips
Dash runs a utility called mksignames on the host to generate the list of
symbolic signal names that can be passed as arguments to the `kill` builtin.
But mips uses a unique mapping of signal numbers, which will likely not match
the build host signal list. So we include a pre-generated signal list for
compilation with mips build targets.

b/22891544

Change-Id: I19540fc3a4ec437d5dc5988b7c8b6b31f71a589a