Skids partial ginstall support.
The Skids kernel has no ARC zImage target; the SDK builds a kernel
image, compresses it with LZMA, and then runs mkimage on that.  The
ginstall change in this commit mimic this behavior.  There is also
some minor refactoring to minimize the number of special cases.

No actual .gi file is generated yet, but the uImage generated will
boot on the Skids bridge.

Separately, there is a small change to google_signing.mk that fixes an
unspecified dependency.  The fix affects current Skids builds because
we don't currently build google_platform (which causes the dependency
in question to get copied).

Change-Id: Iee681f7cabd494582ef9ffcad8423739cfb6033e