Encrypt config partition
Mount config partition on /var/config instead of /config.

Create a filesystem image file in /var/config and mount it with
dm-crypt. The password is wrapped by the TPM.

Move cryptohome init script from S79 to S31 to mount /config ASAP.

Start tcsd in foreground; it will daemonize itself. When tcsd is spawn
in the background, it may not be ready when chapsd and tpm-manager try
to use it. The result is intermittent TPM initialization failure.
(b/23077437)

Change-Id: Iddd466cd15be8f2b943d6760def9446d58871960