google_platform: build wifi utilities
The nl80211.h included in Linux 4.1 (used on the MIPS
platforms like GFHD100 and GFHD200) does not have
definitions for some of the fields we want to use
like 64 bit byte counters. #define a flag to enable
compilation of the extra fields.

Use the C version of wifi_files to replace the Python
version, to lower CPU consumption.

Change-Id: I6a1d7020a0b33dfbea8fa73e11b724580c0250a9