Remove spacecast from is-network-box script
These scripts call is-network-box but has no effect for spacecast:
    S48netmon, S80ifplugd, S99onustats

S99migratemacaddr is a spacecast script but it is obsolete and does not
work. Remove it.

Remove is-network-box check in spacecast version of run-dhclient.

Don't run psstat in S94psstat on spacecast.

Add /config/experiments to silence warnings about read-only /config.

Change-Id: If81e66badc1d07fc60bbaeb764c649b8037c89d7