Make nf conntrack uapi binary compatible w 4.1 kernel
Mindspeed added new values to the two enums ip_conntrack_status and enum
ctattr_type. The problem is that upstream also added new values to these
enums which now overlap with the Mindspeed ones. More specifically,
upstream added IPS_HELPER_BIT = 13 which overlaps with IPS_PERMANENT_BIT
= 13. They also added CTA_MARK_MASK, CTA_LABELS and CTA_LABELS_MASK
which overlap with CTA_COMCERTO_FP_ORIG and CTA_COMCERTO_FP_ORIG.

As a result, we would need different versions of libnetfilter_conntrack
for the old and new kernel. To avoid this, we make the binary interface
of the 3.2 and 4.1 kernel consistent.

Change-Id: I984b3994e3d37b35042187be37290f0f981bbaa9