SpaceCast: Increase min_free_kbytes to 8192
This value is used to indicate the Linux memory management to keep
a minimum amount of memory free. With the default value of 4MB
problems with memory fragementation have been observed when the system
is both serving videos to users and writing data received from the
satellite link to the hard drive. Increasing the threshold to 8MB
aleviates the situation.

Change-Id: Ib23d707ec6e0a81f11c371cb3a58758d0ae24e53
Google-Bug-Id: 24470466