S08harddisk.gfibersc: Omit wait-scsi when booting from hard drive
wait-scsi is normally used to wait for the hard disk to be ready.
wait-scsi checks both ports of the SATA controller and only returns
after a timeout on the second port since no disk is connected. On
SpaceCast there is not need to check for the hard disk to be ready on
a normal boot since the rootfs (and thus S08harddisk itself) is stored
on the hard drive. Only run wait-scsi when the detection of the SATA
drive fails, which might happen on recovery or NFS boots.

Change-Id: I0d67a489293751b5bc01707cf81cd974d949d21e
Google-Bug-Id: 25353130