S06usbinstall.gfibersc: Omit wait-scsi when booting from hard drive
The purpose of the wait-scsi call is to give the system time to
enumerate a USB mass-storage device. If no device is plugged in
wait-scsi returns after a timeout. On a SpaceCast system booting
from the hard drive the system has plenty of time for enumeration
before S06usbinstall is executed. Limit the use of wait-scsi to
non-hard disk boots like recovery and NFS.

Change-Id: Iec46b575a9b353ff4ce67ff0cfa4b65eab71b405
Google-Bug-Id: 25354958