sysvinit: Don't install shutdown.
We don't use controlled shutdowns. The most common cause of
reboot in the fleet is power outage. To avoid giving developers
a false sense of security, we never shut the system down in an
orderly way we just abruptly reset it.

Since /sbin/shutdown isn't used (and doesn't actually work), remove it
altogether.

Change-Id: Ie51264e28a16713fb908de9b2a8e6e1471914687