S40network: fix PCIe enumeration due to fibercl/69562.
fibercl/69562 changes the timing of the kernel's PCIe enumeration such
that the device is not detected at all (previously, the device is
detected but the BAR values have not been initialized). Force a
re-enumeration to detect the Quantenna PCIe device. Because the device
does not already exist, removal is not necessary.

Change-Id: I76f5332e52d987f0e079852152cdf9f0d4327bb0