wireless_tools: increase maxpriv limit.
The maxpriv limit defines the maximum size of an array that is populated
with the private ioctls exported by the interface. Because the array is
dynamically resized through powers of two as entries are added, the
previous limit of 1000 resulted in an effective limit of only 512.

An upgrade of the Quantenna SDK increased the number of ioctls past 512,
which resulted in scripts calling iwpriv to silently fail.

Change-Id: Iba53084ca95a3ac5f05dd2fdf96f3bd0179f4dba