ifplugd:  Fix bug in choosing which interfaces to monitor.
There's a shell script bug in which the wrong variable is checked.
Depending on system state, this either means that too many or not
enough instances of ifplugd may be started.

On my Camaro, $ifc was set and so ifplugd was started for all
interfaces in the list, including the nonexistent wcli1.

Change-Id: I7163f222215de42c75f08987c9020f55a4e07a3d