taxonomy: run dhcpvendortax from dhcp-script
The DHCP Vendor Class provided by many clients can
be used to tell what kind of device it is. This is
particularly useful for VoIP phones and for printers,
where the use of a DHCP Vendor Class is widespread.

Presumably because Enterprise use of these devices
is to put them in specific VLANs, requiring that
the infrastructure be able to identify them.

Change-Id: Ib88df8aea9a6b2c1dc9096840b7238e65ad8fb8b