buildroot: tunes a laser to the discovered channel
This commit adds a persistent userspace daemon that is invoked at boot
time. Every 15 seconds it checks to see if a PON-ID PLOAM message has
been received and was written by the kernel into a sysfs file.

Once the laser channel has been successfully programmed it stores the
current channel in memory so as to avoid continually reprogramming the
laser.

Change-Id: Ie508cf4f33e83e3a1c00e59ad62b8b468077e5ab