Build google_cast for GFHD254
This will not actually enable the google_cast product at this time as
there is no 'cast_shell' binary present for gfhd254. It does mean we
start building the google_cast dependencies, and means that all we have
to do is submit the 'cast_shell' binary once we are ready.

Change-Id: I6c0db6e930d5fae5ee22e3363787dd8f3dc58cfe