Enable mac learning on inband vlan only
With this change, on switch initialization, it ensures ports are
in admin down (shut) to avoid events or packet switching leaking
to CPU while switch is being initialized.
MAC learning is configured as automatic (hardware learning); this
avoids events from going to CPU every time a new SMAC is seen
and allows vlan level learning control.
After this, the switch is configured as usual, and mac learning is
enabled on the inband vlan only.
Ports are then changed to admin up state (no shut).

Change-Id: I2417b810b74f0a1c993e1e371b873a2f1d5edad4