toybox: move df from /sbin to /bin
The recently updated version of toybox moved df from /usr/sbin
to /sbin. Because the way the PATH is ordered, this causes the
init scripts to get the /bin toolbox version, and its output
format leads to parsing errors in S85catawampus.

Fix by moving df to /bin, where it will overwrite the toolbox
version.

Change-Id: I1632431477a14f4ab197028dc28268dac164a92e