s6-rc was forgetting to depend on 'wifi' service.
This resulted in /etc/init.d/wifi not running at startup time, which meant
wlan[01]_portal weren't created, which meant cwmpd didn't create
/InternetGatewayDevice/LANDevice/2/WLANConfiguration/[12], which meant ACS
couldn't configure DHCP (which is a pretty disappointing restriction, but
someone else will deal with that eventually).

We also change wifi to depend on 'network' instead of 'pre-network', which
is pretty heavyweight.  It's only needed because wlan0_portal can't be
created until wlan0 exists, which the 'network' script does.  It would be
better to produce wlan0_portal elsewhere so that this dependency goes away.
Filed b/30485078 for that.

Also, make S40wait-network a "real" init script, to quieten gftest's
smoke/startstop.test.

Change-Id: Ide7b37c9ca3f461ebcf9248214db452e85e0f081