Set fwmark on IPv6 connections to set 802.1q priority
For egress frames on VLAN 2 on wan0, we want to set the Priority code
point (PCP) a.k.a priority bits to 2. For traffic that is software
forwarded by Linux, we achieve this by running the command "vconfig
set_egress_map". However, this configuration is not communicated to the
hardware forwarding engine PFE. For PFE to set the PCP to 2, we need to
set the fwmark field of every connection (TCP and UDP) to 0x8040.

Previously, this has only been done for IPv4 connections. This change is
about adding an ip6tables command to apply the same operation to IPv6
connections.

b/28979744

Change-Id: I1265ef31de25fbbfa2a6fde608015072f5c65a3d