google_oregano-native: Remove dependency on google_cast
We do not yet need to rely on google_cast, and this dependency broke the
gfiber-47 and gfiber-46 builds for CPE 2.5, where Cast is not yet
enabled. Remove the dependency for now; it will be added back in the
future when we do proper libcastcontrol.so integration in
google_oregano-native.

Change-Id: I3905a21b47187d9523f747fa5995492868cc380a