Create and format config fs image file only if TPM is unowned
Previously the init script recreates and formats the config filesystem
image file if it fails to mount it. The old behavior is underterministic
in the rare case that the TPM fails to respond (possibly due to i2c
error). In such case the configuration is erased even though the TPM may
recover after a powercyle. Erasing the config filesystem does not not
help when the TPM fails; a factory reset is required to restore the box.

The new script only creates and formats the config filesystem exactly
once after the tpm-manager has taken ownership of the TPM. Such
condition only occurs during manufacturing or after factory reset.

Bug: 31016860
Change-Id: I764e6ab8c6997c3b00a62cd44ccb4c5b7373bb5b