google_dashplayer: Make the BUILD command specify the build target
Before this CL the buildroot system relies on the fact that the default
target in vendor/google/dashplayer/build happens to work for a BUILD
command. This is fragile - the default target is actually currently
'all', and we only get away with it because the 'install' target
included by 'all' doesn't actually do anything!

Instead, we should specify that we want the 'build' target.

Change-Id: If54a52315d98adb560346f617aa9876f9d571b04