ipapply: disable when using NFS root.
ipapply deletes all IP addresses on wan0 if there is no /tmp/ip/dynamic/wan0
file.  But if we do create such a file, then apman deletes and re-adds the
IP addresses when ifplugd starts, causing NFS to disconnect.

It's too dangerous and frustrating to mess with routes and IP addresses on
wan0 when wan0 is the NFS interface.  Don't do it.

Change-Id: Ia98d09f9a699193d06f6e0feebe9a2a4561e8c95