buildroot: always create the gpon xml file
Create the GPON XML file regardless of whether we have sane hnvram
variables. The factory needs sycl to be running so that there is power
to the SFP for i2c access. If an XML file is missing then it will fail
to start. Additionally, change the placeholder serial number to one that
only includes valid hex characters.

Fixes http://b/31797048

Change-Id: Id81628c70960ef2af231b3f53fb026b1d0e21df3