buildroot:network: tweak tcp_congestion logic
This removes the logic that was changing the TCP congestion control
algorithm for all sockets. Instead, we will only change it for the
sockets that pump media data to the TV boxes - which is done in Sagesrv.

Here, we make sure all algorithms are loaded and made available for
Sagesrv (and other non-root processes) by adding them to the
tcp_allowed_congestion_control list.

Change-Id: Ib182b10e87f71377c0d76c15e1385fc250606ddd