s6-rc: add -O (only) option, and use it in /bin/stop.
Experience shows that nobody actually wanted to stop everything that
depends on, say, 'firewall', when they type 'stop firewall' or 'restart
firewall'.  Give the people (okay, me and gftest) what they want!

As a bonus, this resolves the annoying behaviour of 'restart xyz' where it
would stop all the services depending on xyz, but not restart them again
afterward.

b/31622009

Change-Id: I645cb051dd2026d2663c76b8299d2a8aed6d13c5