Prepare buildroot for building FDS RPM packages
This change generalizes the buildroot google_spacecast package so as to
allow different types of spacecast targets, i.e. appliance vs FDS.
This is required so we can build FDS RPM packages with a properly
versioned compiler.

Config.in: add config option to specify the type of target device
we want to install (appliance or FDS).

fds_defconfig: add minimal buildroot configuration for the FDS.
This builds just the spacecast package and its dependencies,
using the kvm toolchain configured for an x86_64 target.

google_spacecast.mk:
- Remove OUTDIR= argument when calling into spacecast makefiles,
  as they don't use it.
- When installing built spacecast binaries, call directly into the
  spacecast makefile for the desired target device type
  (spacecast or FDS) so these can have independent config rules.
  Also, pass $(BINARIES) to the install rule so that FDS installer
  can copy its RPM package into the out/images directory.
- Update the make clean rules to call the same makefiles as during
  build and install, but obviously with the clean target.

Change-Id: Ia55b02487c44abe8c72636001db86868c6762217