postbuild.sh: add force_version
SpaceCast team are currently cutting their own releases. Since the
person cutting the release is now the same person that creates the tree,
we want to change the build_tagged_spacecast.sh script so that the
release build is a purely readonly operation within an existing
repository. In its current version, the script creates a temporary
spacecast-$version-pvt tag before building in order to force the build
image's version string to the desired value, and removes the tag
immediately afterwards. We would like to provide a different mechanism
so that the desired version string can be passed without having to
write into the existing repository.

Change-Id: I8d957fcd20f33c3aafbd6f8b7d68b4d8ba2dd9c6