Set KBUILD_EXTRA_SYMBOLS in linux.mk for binary kos
If kernel modules that are compiled in-tree reference symbols that are
exported by kernel modules that are compiled out-of-tree, we need to
pass KBUILD_EXTRA_SYMBOLS to the Linux build system.
KBUILD_EXTRA_SYMBOLS needs to be a file path that points to the file
Module.symvers that has been created when the out-of-tree modules were
built.

We currently need this feature to support binary-only kernel modules.
Hopefully, these binary-only kernel modules will disappear in the near
future. The vendor agreed to delivering open-source modules that can be
used in-lieu.

Buildroot packages need to append to the variable
LINUX_KBUILD_EXTRA_SYMBOLS to use this feature.

Change-Id: I6221807ccd1309a115b28a8f097de766b70a2ad5